ml_service/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
Pillow==10.0.0
numpy==1.26.4
torch
git+https://github.com/openai/CLIP.git
